[Fix] Check null whe CheckMonsterClear
Monster가 모두 처치되었는지 체크하는 함수에서 players 리스트의 player가 갖고 있는 model이 null인지 체크해서 null이라면 넘어가는 로직 추가

+++ Assets/PersonalWorkplace/KMS/Script/Boss/BossController.cs
@@ -132,7 +132,7 @@ public class BossController : MonsterController
         foreach (GameObject go in players)
         {
             if (targetEffect) ParticleManager.Instance.GetParticle(effectName, go.transform.position, scale: 1);
-            go.GetComponent<IDamagable>().TakeDamage(Model.BaseModel.finalAttackPower);
+            go.GetComponent<IDamagable>().TakeDamage((float)Model.BaseModel.finalAttackPower);
             DamageText text = DamageTextManager.Instance.Get(go.transform.position);
             text.SetText(BigCurrency.FromBaseAmount(Model.BaseModel.finalAttackPower).ToString());
         }
