+++ Assets/PersonalWorkplace/LGH/Scripts/TestScripts/TestMonster.cs
@@ -4,7 +4,7 @@ public class TestMonster : MonoBehaviour, IDamagable
 {
     private double health = 1000f;
 
-    public void TakeDamage(double amount)
+    public void TakeDamage(float amount)
     {
         health -= amount;
         Debug.Log($"현재 체력 : {health}");
